+++ modules/reporting/__init__.py
@@ -1,0 +1,3 @@
+from modules.reporting.module import ReportingModule
+
+__all__ = ["ReportingModule"]